DB-1367: add separate files with internal and default properties
Internal properties are configured in the
classpath:/application.properties file. Default properties are
specified in /etc/`app`/application-defaults.properties. They can be
overriden in /etc/`app`/application.properties.

The RPM still specifies /etc/`app`/application-defaults.properties as
a configuration file, which may be a problem when the
user (accidentally) modifies this file.