Ignore RRDP delta withdrawal
RPKI objects are automatically removed by the RpkiObjectCleanupJob, so there is no need to remove anything when a delta withdrawal is processed.

Removing the object on withdrawal (or update) should not be done since the object can also be published by another repository(rsync or rrdp) or a malicious RRDP server can fake a withdrawal.