Store manifest objects last
Process the snapshot file twice so that we avoid having manifests in
the store for which some objects have not yet been saved into the
database.

Parsing the snapshot twice keeps memory usage low since there is no
need to keep manifests in memory until the end of the process.