Use depth-first CA validation to reduce memory usage
The code relies a bit less on side-effects and the validation
functions now return a result instead of mutating an accumulator.

All certificate authorities are validated, even when a manifest
contains errors in strict mode. However, the results of the validation
are not kept. This way we avoid having keep child certificates in
memory or to parse the manifest entries twice (first to discover
errors in the manifest, then to process the children).