Only fail the entire manifest when entry hashes do not match
If included objects fail validation, only drop the failed objects, not
all objects on the manifest. The main intent is to prevent MITM
attacks where an attacker replaces or removes objects from the
publication point. Checking that all manifest entries are present and
the hashes match prevents this.

Bugs in the CA or publication server that cause invalid objects to be
published (with the correct hash on the manifest) will not be
prevented and will not cause all objects to become invalid.